Change Floppy to use SATA controller
This commit make Floppy to use SATA controller.
This is required because CoreOS ramdisk which is
used in Ironic doesn't detect floppy devices by default
on VirtualBox.

Also, the file to be attached as Floppy is expected
to be of vdi format now.

It also adds a new mapping for Disk to use SATA controller.